requirements.txt
flask
flask-smorest
python-dotenv
easyocr
Pillow
pytesseract
sqlalchemy
flask-sqlalchemy